FIX(client): Renames file and fixes hotel url  so that it returns a single hotel
