+++ packages/client/src/lib/components/hotelResultsDropdown/hotelResultsDropdown.tsx
@@ -12,7 +12,7 @@ const HotelResultsDropdown: React.FC<HotelResultsDropdownProps> = ({
       {hotels.length ? (
         hotels.map((hotel) => (
           <li key={hotel._id}>
-            <a href={`/hotels/${hotel._id}`} className="dropdown-item">
+            <a href={`/hotel/${hotel._id}`} className="dropdown-item">
               <i className="fa fa-building mr-2"></i>
               {hotel.hotel_name}
             </a>
